extract NonMandatoryRegisters and BoxValue from ergo_box;
