+++ sigma-tree/src/chain/ergo_box.rs
@@ -1,4 +1,8 @@
 //! Ergo box
+
+mod box_value;
+mod register;
+
 use super::token::{TokenAmount, TokenId};
 use crate::{ast::Constant, ergo_tree::ErgoTree};
 use indexmap::IndexSet;
@@ -7,177 +11,13 @@ use serde::{Deserialize, Serialize};
 use sigma_ser::serializer::SerializationError;
 use sigma_ser::serializer::SigmaSerializable;
 use sigma_ser::vlq_encode;
-use std::collections::HashMap;
 use std::convert::TryFrom;
 use std::io;
-use NonMandatoryRegistersError::{InvalidSize, NonDenselyPacked};
 
+pub use box_value::BoxValue;
 #[cfg(test)]
 use proptest_derive::Arbitrary;
-
-/// newtype for additional registers R4 - R9
-#[derive(PartialEq, Eq, Hash, Debug, Clone)]
-#[cfg_attr(feature = "with-serde", derive(Serialize, Deserialize))]
-pub struct NonMandatoryRegisterId(u8);
-
-impl NonMandatoryRegisterId {
-    /// starting index for non-mandatory registers
-    pub const START_INDEX: u8 = 4;
-    /// end index for non-mandatory registers
-    pub const END_INDEX: u8 = 9;
-
-    /// register R4
-    pub const R4: NonMandatoryRegisterId = NonMandatoryRegisterId(4);
-    /// register R5
-    pub const R5: NonMandatoryRegisterId = NonMandatoryRegisterId(5);
-    /// register R6
-    pub const R6: NonMandatoryRegisterId = NonMandatoryRegisterId(6);
-    /// register R7
-    pub const R7: NonMandatoryRegisterId = NonMandatoryRegisterId(7);
-    /// register R8
-    pub const R8: NonMandatoryRegisterId = NonMandatoryRegisterId(8);
-    /// register R9
-    pub const R9: NonMandatoryRegisterId = NonMandatoryRegisterId(9);
-
-    const REG_IDS: [NonMandatoryRegisterId; 6] = [
-        NonMandatoryRegisterId::R4,
-        NonMandatoryRegisterId::R5,
-        NonMandatoryRegisterId::R6,
-        NonMandatoryRegisterId::R7,
-        NonMandatoryRegisterId::R8,
-        NonMandatoryRegisterId::R9,
-    ];
-
-    /// get register by it's index
-    /// `i` is expected to be in range [`START_INDEX`] to [`END_INDEX`] , otherwise panic
-    pub fn get_by_index(i: usize) -> NonMandatoryRegisterId {
-        assert!(
-            i >= NonMandatoryRegisterId::START_INDEX as usize
-                && i <= NonMandatoryRegisterId::END_INDEX as usize
-        );
-        NonMandatoryRegisterId::REG_IDS[i - NonMandatoryRegisterId::START_INDEX as usize].clone()
-    }
-}
-
-/// Stores non-mandatory registers for the box
-#[derive(PartialEq, Eq, Debug, Clone)]
-pub struct NonMandatoryRegisters(Vec<Constant>);
-
-/// Possible errors when building NonMandatoryRegisters
-#[derive(PartialEq, Eq, Debug, Clone)]
-pub enum NonMandatoryRegistersError {
-    /// Set of register has invalid size(maximum [`NonMandatoryRegisters::MAX_SIZE`])
-    InvalidSize(usize),
-    /// Set of non-mandatory indexes are not densely packed
-    NonDenselyPacked(u8),
-}
-
-impl NonMandatoryRegistersError {
-    /// get detailed error message
-    pub fn error_msg(&self) -> String {
-        match self {
-            InvalidSize(size) => format!(
-                "invalid non-mandatory registers size {} (expected {})",
-                size,
-                NonMandatoryRegisters::MAX_SIZE
-            ),
-            NonDenselyPacked(reg_id) => format!(
-                "non-mandatory registers are not densely packed, {} is missing in range [{} .. {}]",
-                reg_id,
-                NonMandatoryRegisterId::START_INDEX,
-                NonMandatoryRegisterId::END_INDEX
-            ),
-        }
-    }
-}
-
-impl NonMandatoryRegisters {
-    /// Maximum number of non-mandatory registers
-    pub const MAX_SIZE: usize = 6;
-
-    /// Empty non-mandatory registers
-    pub fn empty() -> NonMandatoryRegisters {
-        NonMandatoryRegisters(vec![])
-    }
-
-    /// Create new from map
-    pub fn new(
-        _regs: HashMap<NonMandatoryRegisterId, Box<Constant>>,
-    ) -> Result<NonMandatoryRegisters, NonMandatoryRegistersError> {
-        // return error if size is incorrect and/or there is a gap
-        // we assume non-mandatory indexes are densely packed from startingNonMandatoryIndex
-        // this convention allows to save 1 byte for each register
-        todo!()
-    }
-
-    /// Create new from ordered values (first element will be R4, and so on)
-    pub fn from_ordered_values(
-        values: Vec<Constant>,
-    ) -> Result<NonMandatoryRegisters, NonMandatoryRegistersError> {
-        if values.len() > NonMandatoryRegisters::MAX_SIZE {
-            Err(NonMandatoryRegistersError::InvalidSize(values.len()))
-        } else {
-            Ok(NonMandatoryRegisters(values))
-        }
-    }
-
-    /// Size of non-mandatory registers set
-    pub fn len(&self) -> u8 {
-        self.0.len() as u8
-    }
-
-    /// Return true if non-mandatory registers set is empty
-    pub fn is_empty(&self) -> bool {
-        self.0.is_empty()
-    }
-
-    /// Get register value
-    pub fn get(&self, _reg_id: &NonMandatoryRegisterId) -> Option<Box<Constant>> {
-        todo!()
-    }
-
-    /// Get ordered register values (first is R4, and so on, up to R9)
-    pub fn get_ordered_values(&self) -> Vec<Constant> {
-        self.0.clone()
-    }
-}
-
-impl From<NonMandatoryRegistersError> for SerializationError {
-    fn from(error: NonMandatoryRegistersError) -> Self {
-        SerializationError::Misc(error.error_msg())
-    }
-}
-
-/// Box value
-#[derive(PartialEq, Eq, Hash, Debug, Clone)]
-#[cfg_attr(feature = "with-serde", derive(Serialize, Deserialize))]
-pub struct BoxValue(u64);
-
-impl BoxValue {
-    /// Create new value (with bounds check)
-    pub fn new(v: u64) -> Option<BoxValue> {
-        if BoxValue::within_bounds(v) {
-            Some(BoxValue(v))
-        } else {
-            None
-        }
-    }
-
-    /// Check if a value is in bounds
-    pub fn within_bounds(v: u64) -> bool {
-        v >= 1 && v <= i64::MAX as u64
-    }
-}
-
-impl SigmaSerializable for BoxValue {
-    fn sigma_serialize<W: vlq_encode::WriteSigmaVlqExt>(&self, w: &mut W) -> Result<(), io::Error> {
-        w.put_u64(self.0)
-    }
-    fn sigma_parse<R: vlq_encode::ReadSigmaVlqExt>(r: &mut R) -> Result<Self, SerializationError> {
-        let v = r.get_u64()?;
-        Ok(BoxValue(v))
-    }
-}
+use register::NonMandatoryRegisters;
 
 /// Transaction id (ModifierId in sigmastate)
 #[derive(PartialEq, Eq, Hash, Debug, Clone)]
@@ -375,15 +215,6 @@ mod tests {
     use proptest::{arbitrary::Arbitrary, collection::vec, prelude::*};
     use sigma_ser::test_helpers::*;
 
-    impl Arbitrary for BoxValue {
-        type Parameters = ();
-        type Strategy = BoxedStrategy<Self>;
-        fn arbitrary_with(_args: Self::Parameters) -> Self::Strategy {
-            // TODO: should be in 1 - i64.max range
-            any::<u64>().prop_map(|v| BoxValue(v)).boxed()
-        }
-    }
-
     impl Arbitrary for ErgoBoxCandidate {
         type Parameters = ();
 
@@ -393,15 +224,14 @@ mod tests {
                 any::<ErgoTree>(),
                 vec(any::<TokenAmount>(), 0..10),
                 any::<u32>(),
-                vec(any::<Constant>(), 0..7),
+                any::<NonMandatoryRegisters>(),
             )
                 .prop_map(
-                    |(value, ergo_tree, tokens, creation_height, constants)| Self {
+                    |(value, ergo_tree, tokens, creation_height, additional_registers)| Self {
                         value,
                         ergo_tree,
                         tokens,
-                        additional_registers: NonMandatoryRegisters::from_ordered_values(constants)
-                            .expect("error building registers"),
+                        additional_registers,
                         creation_height,
                     },
                 )
